unify RestClient to one common class
